+++ Backend/package.json
@@ -12,7 +12,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@auth/core": "^0.19.0",
+    "@auth/core": "^0.19.1",
     "@xenova/transformers": "^2.10.0",
     "cookie-parser": "^1.4.6",
     "express": "^4.18.2",
